1984-refactor-tests
- Implement pending tests
- Remove duplication in tests
